Add optional name filter
Add a name filter to help out sending input messages to multiple accessories.
The filter is only activated if a msg.payload.name is present. The name should
be exactly the same as the one sent outgoing messages.
I the msg.payload.name is not present then the message is handled like normal.